+++ server/sonar-server/src/main/resources/org/sonar/server/computation/ws/project-example.json
