+++ server/sonar-server/src/main/resources/org/sonar/server/computation/ws/queue-example.json
